refactor: move unit tests
