+++ lua/gitsigns.lua
@@ -36,9 +36,7 @@ local M = {}
 
 
 
-if _TEST then
-   M._tests = {}
-end
+
 
 
 M.detach_all = function()
@@ -87,15 +85,6 @@ local function parse_fugitive_uri(name)
       name = root_path .. sub_module_path .. real_path
    end
    return name, commit
-end
-
-if _TEST then
-   M._tests.test_name_parse = function()
-      local path, commit = parse_fugitive_uri(
-      'fugitive:///home/path/to/project/.git//1b441b947c4bc9a59db428f229456619051dd133/subfolder/to/a/file.txt')
-      assert(path == '/home/path/to/project/subfolder/to/a/file.txt', string.format('GOT %s', path))
-      assert(commit == '1b441b947c4bc9a59db428f229456619051dd133', string.format('GOT %s', commit))
-   end
 end
 
 local function get_buf_path(bufnr)
@@ -458,6 +447,10 @@ M.setup = void(function(cfg)
 
    autocmd('DirChanged', debounce_trailing(100, manager.update_cwd_head))
 end)
+
+if _TEST then
+   M.parse_fugitive_uri = parse_fugitive_uri
+end
 
 return setmetatable(M, {
    __index = function(_, f)
