+++ teal/gitsigns/util.tl
@@ -2,12 +2,6 @@ local record M
   type FmtInfo = {string:string|integer|{string}}
 
   path_sep: string
-
-  _tests: {string:function}
-end
-
-if _TEST then
-  M._tests = {}
 end
 
 function M.path_exists(path: string): boolean
@@ -186,13 +180,6 @@ function M.expand_format(fmt: string, info: M.FmtInfo, reltime: boolean): string
 
   ret[#ret+1] = fmt
   return table.concat(ret, '')
-end
-
-if _TEST then
-  M._tests.expand_format = function()
-    assert('hello % world % 2021' == M.expand_format('<var1> % <var2> % <var_time:%Y>' , {
-      var1 = 'hello', var2 = 'world', var_time = 1616838297 }))
-  end
 end
 
 return M
